requirements.txt
google-genai>=0.2.0
torch>=2.0.0
numpy>=1.24.0
Pillow>=10.0.0

